+++ src/structures/mod.rs
@@ -6,10 +6,3 @@ pub mod version_structs;
 pub type Datetime = chrono::DateTime<chrono::Utc>;
 /// A base 62 number stored as a string
 pub type ID = String;
-
-#[derive(serde::Deserialize, serde::Serialize, Debug, Clone, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum ProjectType {
-    Mod,
-    Modpack,
-}